Added Super Smash Bros 2DS v1.1.7 Plugin
Releasing a NTR Mode 3 Cheat Plugin made for Super Smash Bros 3DS for the Old2DS. Includes cheats for VS. Mode, Classic Mode, and A 0% Damage Cheat for Smash Run. Also includes a character modifier cheat, but it only works to allow the Miis from the Fighting Mii Team in as extras.