Added All OS TEST Versions of Randomizer Tool
This is in response to the Randomizer Tool not working for all Operating Systems and Processors. This is by no means a finished version of these tools, only a version to be tested and refined later. After testing, if these work for several new Operating Systems and Processors, this tool will be the next/final version of the tools. These tools attempt to tackle the previously mentioned issue by scanning for a signature in the memory and then writing to memory in areas relative the signature. The source code will be uploaded after it reaches the final version.